test: split profile transfer integration suites
